requirements.txt
azure-functions
azure-storage-blob
paramiko